loader/requirements.txt
qdrant-client
PyPDF2
python-docx
markdown
requests
psycopg2-binary